Enhance Docker and HTTP server configurations
- Added a stop grace period of 120 seconds to the Docker Compose configuration for the lotusd service to ensure a smoother shutdown process.
- Introduced a new Dockerfile for prebuilt images, setting up the environment with necessary dependencies and configurations for the Lotus application.
- Refactored the HTTP server implementation to utilize cpp-httplib, improving the handling of HTTP requests and responses.
- Updated logging throughout the HTTP server and related components for better clarity and consistency.

These changes aim to improve the deployment and operational stability of the Lotus application.

Made-with: Cursor

# Conflicts:
#	src/init.cpp
#	src/validation.cpp

+++ src/uint256.h
@@ -88,7 +88,7 @@ public:
 
     const uint8_t *end() const { return &m_data[WIDTH]; }
 
-    unsigned int size() const { return sizeof(m_data); }
+    size_t size() const { return sizeof(m_data); }
 
     uint64_t GetUint64(int pos) const {
         const uint8_t *ptr = m_data + pos * 8;
